Add temporary views for Cpu/MainMemoryProfilerStage
For the sake of getting the MVP of the Task-Based UX checked-in
quicker, two new temporary views for the CpuProfilerStage
and MainMemoryProfilerStage have been checked-in. No existing
compose code existed before for these views, and with Jewel
coming in and risking refactoring/rewriting to this UI code, it
makes sense to save the refactor for later and unblock ourselves
from this end now. Nonetheless, a tracking bug was added to each
class for doing the inevitable update of these views to be in
Compose + changing the name to something more descriptive.

Bug: 277798877
Bug: 277798936
Test: manual
Change-Id: I4215fb6f700770440e7f4eca1480440c10ce2014