Add dropdown to pick AGP version in dev Studio
No user-visible change, this is only visible when idea.is.internal=true

This currently doesn't work well with very old AGP versions as some of
the templates need to be more parameterized by version to work well.

Also removed a fallback case that tries to determine the AGP version by
reading all build files when trying to add a new module to a project
without valid Android models.

Bug: N/A
Test: Some automated unit tests, some manual
Change-Id: I6d0de24ebbba7479903b0c80b4d8ff9a680ff73f