Fix res package resolution in non-canonical order
Adds support for correctly resolving resource references of the format
`?attr/android:attribute`. (This has the resource type and package
swapped in order from the canonical `?android:attr/attribute`.) While
this isn't documented publicly, aapt allows this format. Additionally,
there are enough usages in google3 that it was decided in b/112570942 to
maintain minimal editor support for the format.

This follows that pattern - resource reference parsing will recognize
and correctly resolve this format, even though we won't suggest or
otherwise promote it.

Fixes: 296217029
Test: AndroidNamespacedValueResourcesDomTest
Change-Id: If577ba813e40f668d7183b31cf2738ff8fe40b53