Add a run-time assertion
The PropertiesFiles utility refreshes the VFS if necessary in order to
assure consistency, as some clients are working with raw files
primarily, and in order for those changes to be reflected in the VFS
we need a refresh.

This refresh must not be performed with a read lock, as that would
deadlock.  However, inserting the assertion that the read lock is not
held reveals that we do in fact call this routine with the read lock
held in various places; we do not observe this deadlock in practice,
at least not much, because usually the refresh is not needed (the VFS
is up to date at the point where it is queried).

However, each such call is a potential deadlock, so insert the
assertion and fix the callers not to call PropertiesFiles (even
indirectly) with the read lock.

Bug: 176220349
Test: existing
Change-Id: I025981a416bdbe3ee67a45bbb06731c2621eaa60