Refactor DeviceManagerConnection to allow testing
Instead of deferring initialization of the DeviceManager, take it as a
constructor argument. Every caller that gets a DeviceManagerConnection
immediately invokes a method that would force initialization, so this
doesn't change the timing of initialization, and initIfNecessary will
fail if and only if sdkPath is null, so it can't start working later.

This allows us to use a temporary DeviceManager that gets disposed at
the end of the test, rather than the singleton DeviceManager cache
shared with other tests.

Add a test that createDevices avoids name collisions.

Bug: 299189247
Test: Added
Change-Id: I862f7a28e90bdf38fa45243fe078c4da02661bff