Send typing reports when Essential Mode changes
The impacts of Essential Highlighting on typing latency are still being
analyzed. To help with this, send out reports right when the state
changes. This makes sure reports are purely attributed to which mode
they were in. Currently, analysis has to be done on whether or not they
were in the mode part time or not and if they were, the report has to be
discarded as we don't know which typing can be attributed to which mode.

This had to wait for IDEA-330383 to be fixed and added with ag/24680566

Fixes: 292003320
Change-Id: Ib16fd570e823b0f1cced738272792c3b3f5fbfa6