Simplify UnusedResourceHandler.invoke APIs
In every place that `invoke` is called, the last two parameters always
have the same value as each other. This change coalesces them into a
single value, and then splits them into two separately-named functions
that don't take the boolean parameter(s) in the first place.

Additionally, the existing `skipIds` parameter was poorly named and
actually did the opposite of what it suggested. I've verified that the
non-test usages of this parameter are functionally correct with the way
the existing behavior worked, so it was just the naming that was wrong.

Bug: 297043493
Test: UnusedResourcesTest, UnusedResourcesGradleTest
Change-Id: I2be1ea77eb40713964bd19005c23a0821aa357aa