Store quick boot targets for running devices
If an AVD is running when the Select Multiple Devices dialog is opened,
and the user selects that AVD, currently it will only be persisted as a
RunningDeviceTarget, which means that when it goes offline, it will be
removed from the set of selected devices.

While we don't know how the device was booted, we can default to quick
boot, so that the user's intent to target that device is not lost; in
the relatively rare case that the user needs a different boot target, it
can be selected when the device is offline.

Fixes: 238858112
Test: Added
Change-Id: If65e5874d1583862f5231cc62422e9b223e84f91