Update BindingXmlIndex to avoid parent dir name
Parent directory name cannot be accessed at indexing time, per the
contract of `FileBasedIndexExtension`. Moving logic around so that we
perform the check based on directory to the time when we access the
index, rather than when we're building it.

This means that we may hit false positives and over-index xml files. I'm
introducing a minimum heuristic to avoid indexing totally unrelated
files. I haven't been able to identify any other good heuristics, but
can look into it further if this proves to be a performance problem (I
don't anticipate that it will be, since in the false positive case the
amount of data stored will be minimal.)

Fixes: 230115126
Bug: 204421590
Test: BindingXmlIndexTest
Change-Id: I1984dfc458662161b44b8d0d02dd60c30efa96c3