Listen until next sync if its state is UNKNOWN
When opening a project the sync status could be of UNKNOWN type.
When this happens, the code split and design panes aren't shown.

The UNKNOWN state is the very first status of when a project is opened.
When on this status the sync of the project is not started yet.

This change will listen till the next sync not only if the sync status
is in progress but also when it is in an UNKNOWN state.

Fixes: 299881938
Test: SyncUtilTest, SourceCodeEditorProviderTest
Change-Id: I0318ad36500fdb93863a3653546de1a3d8f8debb