Always switch to a new preferred process
We should not keep the old process if that happens to be a process
with the same process name.

This change would fix a connect issue if the user restarted an app
and the transport/app inspection for some reason would send new
process before a reporting the old process dead.

If that happened this code would keep the Layout Inspector connected
to the old process (which is not possible).

Fixes: 300285134
Test: ProcessesModelTest
Change-Id: I1b89509fbb7527d3df45040ee33c7032c7011bf5