Indirect Gradle daemon memory settings UI through a token
The daemon memory settings are inherently project-system specific.
Solve the fact that we need to display and modify
project-system-specific information in a generic UI control by adding
a layer of indirection, through a project-system token which is
responsible for creating and managing the project-system-specific
portion of the UI.

This should fix the issue that Gradle-related UI is presented to ASwB
users, and also should allow us to move of a substantial chunk of
com.android.tools.idea.gradle out of android core.

Bug: 147586218
Bug: 299469874
Bug: 300434824
Test: existing, manual
Change-Id: I57fb119ff7eec444ccf7dad4a573ada78898cdd3