Compute project memory setting recommendation through token
The project system needs to be involved, because the project
configuration can only be used to infer other aspects (such as
expected memory needs) if there is knowledge of how the details of
the project map to the external world.

Although the existing code looks generic, simply counting the number
of modules in the project, in fact that is highly project-system
specific; Gradle generates four (IDE) modules per (Gradle) project in
the system, while a Blaze project currently has exactly one module for code, and an arbitrary number of modules for library resources.

Adjust the tests to have a minimal Gradle-based project structure,
rather than a light project.  We need to mock ModuleManager quite
tightly in the tests, in order that the tearDown methods (which want
to use the *real* ModuleManager) do not cause spurious errors.

Bug: 301236834
Test: existing
Change-Id: I9f8849f9899c22dc405320a159a0169ae99a2341