Make sure we properly check unit tests are re-executed
The test function testTasksIsReExecuted() was not really
testing that the test is re-executed for the appropriate
reason. We should be re-executing unit tests tasks because
we do set Task.UpToDateWhen to false in the tests init scripts.
In this function we were asserting that the tests tasks were
in fact being re-executed but the reason behind that was
actually because we were using the same GradleExecutionSettings
object for the 2 invocation, and this is wrong because
GradleExecutionSettings is specific to every invocation. What
happened in this test is that the GradleExecutionSettings
would get modified during the first task execution, and then
when invoking the tests tasks for the second time,
Gradle would detect that the settings had changed and then
re-execute the tests tasks for the wrong reason.

Bug: N/A
Test: this
Change-Id: Id9c62395d31d5d8111371dff46982df2e68a39b8