Add golden dir to structure of bazel outputs
When running TemplateDiffTest in GENERATE_GOLDEN=true mode, instead of
dumping all the generated files in the top level directory that gets
zipped into outputs.zip, put them in a golden/ directory. This makes the
command to unzip to source simpler because we don't need the extra line
to delete idea.log anymore. It also makes room for other files (Lint
baseline) to be included in outputs.zip in the future.

Bug: 272751498
Test: this is the test

Change-Id: I84e747fbcf3d8e5f9c95c5b87f5c20c9bd951472