Move existing deployment combobx to legacyselector
Add a flag to switch between the old and new implementations of the
deployment target dropdown, in the legacyselector and selector packages.

Add delegates for extension points and services declared in plugin.xml
that select the implementation based on the flag.

For actions, use DeploymentTargetSelectorActionsInitializer to replace
old actions with the new version if the flag is enabled.

Bug: 299536545
Test: existing unit tests preserved for old code; manually verified flag
Change-Id: Ib851c0997f550e6a0df6108300727b0725c191e2