Return a GoogleMavenArtifactId for item getGradleCoordinateId
We should only ever be dealing with artifacts under our control.  This
is now (a little bit more) manifest in that DependencyManager in this
module now deals with GoogleMavenArtifactId objects rather than bare
Strings.

This also allows removing the empty-string-literal IN_PLATFORM
interned String, with its potential for looking like a constant with
an arbitrary value but in fact depending on its exact form.

Bug: 147586218
Test: existing tests adjusted
Change-Id: I7dd75f00c2b7c1c518e596d4e216502a1e32f47c