Handle StatusRuntimeException
In the interval of time between when an emulator is stopped and
Studio detects it stopped, it's possible for clients of the
Transport to send messages to that emulator. Doing so causes
"StatusRuntimeException: UNAVAILABLE: Network closed for unknown
reason" to be thrown by grpc.

By catching the exception here we prevent it being thrown by grpc.

Test: NA
Bug: 302320067
Change-Id: I73378784611f678fc3cf4a76f274b3a35265cc22