Generate artifact resolvers through a project system token
Existing tests are minimally adjusted but this change otherwise leaves
the logic alone, in order to help make it clear that this is a code
reorganization with effectively no change in production functionality.
A followup CL will remove the vestigial Blaze support (to be replaced
by a cleaner implementation) and remove interfaces that are no longer
used.

Bug: 279762135
Test: existing tests adjusted
Change-Id: Ib2751501f7d8d4b9867cb06ab7c4b40aefdf9c50