Add suppress quick fix for issues in UI Check mode
This adds a way to suppress Visual Lint issues found in UI Check mode
through a quick fix. When the suppress action is triggered, it adds a
new @Suppress annotation on the Composable that started the UI Check
mode. As an argument to this annotation, it uses the type of the error,
so that errors of other types can still be found.
Suppressing several kinds of issues is possible with this, by adding
several types to the annotation arguments.

Fixes: 302196745
Test: test added
Change-Id: I9e62a500352aa9b506bd79504c100e2a4ff3b719