Replace Range constructor in CPU artifact fetching
The Range constructor used in the CpuCaptureSessionArtifact's
getSessionArtifacts method is currently invoking the Range aspect
as a side effect. This performs an uncessary aspect firing and
handling on a repeated basis, as getSessionArtifacts is called
by the updater. This constructor has been replaced with the one
that takes in the range values and does not have any side-effects.

Bug: none
Test: existing
Change-Id: Ice5740320c3e29eed36a180aa65ffdf4d1dab89e