Shows alpha colors in resource manager
Resource manager didn't parse the alpha color when changing colors
in the colors.xml file. The issue was caused by the color parser having
the alpha disabled.

A customised Color.toHex function needed to be implemented as the format
of java.awt.Color is #RGBA but we needed to show #ARGB as we define
in our resources.

Fixes: 278844165
Test: AssetDataProviderTest
Change-Id: I19fb881a22e52ccb23f4cadcbf32fe5ccb28ab18