Aggregate mouse scroll wheel events between renders
...so that not every scroll wheel event fires off a model
change. Rather, we can collect all mouse wheel events made
in between renders, squash their values together into one
aggregate value, and then update the model once per render.

Because all instances of the TimelineScrollbar utilize the
StreamingTimeline implementation of the Timeline interface,
the TimelineScrollbar constructor was modified to take in
a StreamingTimeline instead of a Timeline. This allows us
to avoid having to check at runtime if the Timeline object
was of the StreamingTimeline implementation everytime we
want to modify the mouse scroll event queue that is
exclusive to the StreamingTimeline from the scrollbar class.

Context: The problem identified as the cause of freezing/lag
when scrolling through a streaming timeline was that the mouse
wheel events were firing many times between renders, and each
event was updating the timeline model, despite the actual change
not being visualized until the next render. Thus, as stated
earlier, we now will aggregate all the mouse wheel events
between renders, and will only perform maximally one model
update per render.

Fixes: 279091246
Test: existing tests updated + manual stress test performed
Change-Id: If49a6b3851d07507380c8f835ae81c37cb8735d2