[android-kotlin] Combine logic for component registration intentions.
These intentions all register an Android component (activity, service,
or broadcast receiver) in the manifest. Note that this is the only
intention that does K1/K2 switching inline. This intention should
eventually be converted to UAST and combined with the Java
implementation of the same intention, to minimize duplicate effort.

Bug: 284252165
Test: Treehugger
Change-Id: I143cfbc325fef2e970c4d68892b9cf264f6f035e