Log delegate action if action is AnActionWrapper
Actions can be wrapped in AnActionWrapper to modify their behaviours.
However, the health monitor logs the class name for the action
performed, which in the case of action wrappers will be the name of the
wrapper. That conflates several different actions under the same name,
and hides the important information of what the underlying action is.
This changes the health monitor to log the delegate action class name
when dealing with action wrappers.

Bug: N/A
Test: N/A
Change-Id: I957a0bcb5750460c2e0171dd14604f7bb6f368c9