Fix flaky PreviewRefreshManagerTest
All test scenarios in this file are built on top of the premise that
we can control the enqueueing order of refresh requests. But the
requestRefresh method executes the enqueueing asynchronously, meaning
that some race conditions were affecting the assumptions made by the
tests.

This change exposes a TestOnly method in RefreshManager to allow
synchronizing the enqueueing of requests.

Fixes: 303196197
Test: executed with --runs_per_test=999
Change-Id: Ie81b543107120aa764bdb4b5956764532aafbef8