Cache gradle files in AndroidBuildScriptsGroupNode
This should help avoiding expensive files set computation on each
'contains' call. Update this cache when children are updated. Not
sure if there is an actual way to trigger refresh of this node's
children list, but it seems to be the right thing to do just in case
this changes in the future.

Fixes: 215570559
Test: Added test case to AndroidBuildScriptsGroupNodeTest
Change-Id: Id74ee9b150e0321975d0f4caa87a10368576544d