Use generic sync manager in AndroidNewModuleAction
The rest of it is written generically.  This change might alter
behaviour slightly in non-Gradle-based projects, where this action
will be disabled while the project is being (non-Gradle-)synced,
whereas previously if the project system allows file creation it would
always be enabled; however, since this action should probably not be
visible for non-Gradle projects, if any change is visible it is in the
correct direction.

Bug: 305009551
Test: existing
Change-Id: Ie6b58907dea94a216507c5a838d0f974e4b1d1e1