Move invisible previews out of the surface
When a scene view is invisible, the various layout managers for Compose
preview set its position at (-1, -1). But since the scene view has a
non-zero width and height, part of it is still present on the main
surface, even if invisible. When calling DesignSurface.getSceneViewAt,
we can potentially get an invisible scene view in return. To avoid this,
layout managers set the position of invisible preview to
(Integer.MIN_VALUE, Integer.MIN_VALUE).

Bug: 305187385
Test: Checked manually
Change-Id: Ifd793b2eb55b08b5b6fa591ea05b1f44e4275a67