Add task termination logic
Because tasks and sessions have equivalent lifecycles under
the Task-Based UX, when the task is considered to be finished,
the correspnding session should be terminated. Thus, this
change adjusts the commands used to stop recordings underlying
a task so that the daemon, when handling the command, not
only executes the termination of the task/recording, but also
ends the current session the recording is under.

Bug: 277797528
Test: updated + added
Change-Id: I92991444a91a1fc191953f44f547156a54d46b5c