Slightly more correct isGradleModule()
For the deprecated AndroidModuleInfoProvider, provide something which
actually differentiates per-module in theory.

It's unclear to me whether the use of this AndroidModuleInfoProvider
is safe, in that it seems to conflate "non-Gradle" with "legacy JPS",
whereas there might be other build systems, now or in the future.  I
suppose there's a race between this being noticeable and
AndroidModuleInfoProvider being deleted, but in the intervening time
this change should not result in any visible change in behaviour.

Bug: 305708611
Test: existing
Change-Id: Ic22a3206da686efbf8b031102492f7d14b61a5c4