Mitigate leak of WorkBench
IntelliJ does sometimes retain disposed editors via the
DaemonCodeAnalyzer. This, when pointing to a preview panel can cause the
leak of a large number of objects.
WorkBench was doing removeAll but that did not remove the pointers to
all the existing panels. This change forces the removal of all panels in
the WorkBench on dispose.

Test: Existing
Bug: N/A
Change-Id: Ia4b3685143ec49089f79d328e8b679a5bd8458c4