Replace uses of isBuildWithGradle in test run configurations
Replacing them with a test on whether the project contains
GradleFacets does not actually solve the problem that they are
committing modularity violations.  In one case, the modularity
violation is practically enforced by the platform, where we have
possibly the most egregious example of emulating double-dispatch in a
message-passing object system I have ever encountered (see
isPreferredTo / shouldReplace in ConfigurationFromContext); in the
other, the Gradle-ness or otherwise is being used to determine whether
to make a dialog field editable.  We will deal with both instances at
a later date, but for now this change means that they no longer call
the surprisingly heavyweight isBuildWithGradle() method.

Bug: 305708611
Test: existing
Change-Id: I85b7af96bbb08975784d989a23f8af66bf15c496