Do not trigger sync on non-android KMP source sets
The `AndroidGradleProjectStartupActivity` validates the integrity of
Android models and models when a project is opened and triggers Gradle
sync in the case of any inconsistencies, which most likely are results
of previously misbehaving sync.

(1) Do not validate non-Android modules in KMP module groups even when
    and android target is present.

(2) Do not trigger sync when a library without any entries is present
    in the project. The original intention of the existing check is to
    trigger sync when artifacts expire in the Gradle cache.

(3) Use a map to lookup well-known android source sets.

(4) Add a test to prevent regressions.

Bug: 305944782
Test: OpenProjectIntegrationTest
Change-Id: I7df6e689bc91d0fbfd3fe1a750d46f4fd263ca13