Don't use cached models across Studio versions
The current constant is sometimes forgotten, and that leads to difficult
to diagnose broken behavior at runtime if the model classes are reloaded
with unexpected missing data.

Instead, always force a resync when running with a different version of
Studio.

This change has no effect on development builds, as they already all
have the same version.

Bug: N/A
Test: Manual
Change-Id: I349d553258759abc627884ab6381a982cd911727