Customize WarningLayer for UI Check mode
For UI Check mode, we want the layer to react to issue selection by
animating a border around the affected components. This border is a
gradient and its width diminishes over time.
In addition, for UI Check mode, we do not want to highlight the contour
of the entire preview.

Bug: 287121972
Bug: 287311183
Test: Cehcked animation manually
Change-Id: I3f155622c769971ce79b4e5bf51835b4c00c3b23