Capture screenshots on Windows
When testing on Windows, our test environment does not have a GUI, so
there's no way to capture the entire screen. As a result, it's very
difficult to diagnose test failures.

This change allows the framework to capture screenshots of every Swing
window every second, then it makes a best-effort attempt to stitch them
into a video at the end.

Fixes: 253452847
Test: manual (it involves inspecting images/videos)
Change-Id: I3cfcf6e152e586b3cc93f9911fc1feac05cd1b26