Improve reporting of unknown sync failure reasons
If failure reason value was not collected properly then by looking
at exception try to figure out if it was treated by our or platform
code.
This should help us to understand if we report all the values we can
properly and will help us detect if any new errors remain unreported
by mistake in the future.

Bug: 306405462
Test: create SyncFailureUsageReporterTest
Change-Id: I929054718dd83805ff23ce1152e2914bb3b1d484