Ensure recomposer is cleaned up on dispose
This is to ensure that there are no leaked compositions. Additionally,
enable the hot-reload mode in the Compose Runtime since there is some
additional error handling that allows for exceptions to happen within
the Composable.
Preview will throw exceptions for example when the Preview is
interrupted. This tries to ensure that the compositions are correctly
disposed.

The dispose clean-up will ensure that there are no instances of
androidx.compose.runtime.Recomposer$RecomposerInfoImpl remaining after
the render.

Alternatively, we manually dispose any pending recompositions when the
RenderTask is disposed.

Fixes: 301314971
Fixes: 179195773
Test: SingleComposePreviewElementRendererTest
Change-Id: Id8dbd948e764d9eefe64e14c00f31a449f54bf22