Add TaskGrid component to display tasks to select
This component will be a major piece of the home tab, allowing
the user to see what tasks are available post process selection.
The user can select one of the displayed, available tasks to
pursue. As of now, the task grid has been implemented for only
the home tab as it takes in the selected process + device. But
much of the code will be shared with a similar overloaded
implementation for past recordings that will be focused on the
selected recording, which has a slightly different UI.

Bug: 277798754
Test: added
Change-Id: I85c87bd7c592490036c91fbcb7d6791f5c527287