Set the owner to shell for files added to /data/local/tmp
Explicitly set the owner of the perfd folder and files uploaded
to the perfd folder in case the user has temporarily rooted the
device. With this code in place the owners of the uploaded files
are set correctly such that transport will work if the user
unroots the device later.

This change does not attempt to fix a bad situation created by
an earlier version of Studio.

Fixes: 301614323
Test: TransportFileManagerTest
Change-Id: Ic2c6882db4e2aeef12d4633f1c66a7c52ef794be