Fix ordering of ComposeCompletionContributor
The "Turbo Complete" KindExecutingCompletionContributor is somehow
caching and returning previous results. From what I can tell, it's an
effort to provide faster completion results. However, it's running prior
to our contributor, so we don't have a change to decorate the results.

Changing the ordering fixes this. Longer-term there may be a way to more
formally plug into this infrastructure, but since this still seems to be
experimental the ordering should be fine for the time being.

Fixes: 303847909
Test: Manually verified
Change-Id: Id5a889cc02dee9bf80be06a0a653152a8d442695