Reimplement Device and Target
Device is now a concrete class that delegates to
DeviceProvisionerAndroidDevice; we don't need subclasses because the
differences between device classes are handled by the Device
Provisioner.

Target has two main structural differences: first, it has no subclasses
for the different modes of booting; it simply takes BootOption as a
parameter.

The bigger change is that it holds a Device rather than a Key
(DeviceId). This avoids the need elsewhere in the code to look up the
current Device from the Key, but it also means that we can't directly
deserialize into a Device. Instead, we deserialize to a TargetId,
then resolve that to a Target based on the current device list.

Bug: 299536545
Test: added
Change-Id: Ic34ffd26760dbe8de050972d0b9d087d995c6be5