Adapt to Jewel 0.8.1 upgrade
...as Jewel 0.8.1 brought some dependency restructuring.

Part of this change includes separating the "jewel.xml"
library into two libraries: one for standalone use
("jewel-standalone.xml") and one for ide-plugin use
("jewel-ide.xml"). The jewel-standalone library should be
treated as an add-on for the jewel-ide library, as the
jewel-ide library contains all the foundational dependencies
for jewel and can support jewel development in isolation,
unlike jewel-standalone.

Bug: 272822954
Test: updated
Change-Id: Ifa5745717864f6bc68d87a8b6cc6d3c87d914c12