Reimplement DevicesSelectedService
The DevicesSelectedService is now the central coordination point for the
target selection. It consumes the flow of present devices from the
DevicesService (upcoming) and provides a DevicesAndTargets object
containing a consistent view of the present devices and the selected
targets.

It receives commands from the UI to update the selection, and persists
the state accordingly.

Bug: 299536545
Test: Included
Change-Id: I93422b7f5906fbbad6480f562e1d7e8404039cf5