Create a separate toolbar for embedded Layout Inspector
Accoring to the mocks, the toolbar for embedded Layout Inspector
should contain a label with "Layout Inspector" on the left, and
actions on the right. The actions should include a gear action
that when clicked shows a menu of actions.

This CL hides the gear and collapse action from the tool window
definitions and implements the new toolbar for embedded LI.

Test: NA
Bug: 304590546
Change-Id: I8247cded4869af81a0ae8a689dc83614771a5d0a