Support multi event in Crashlytics
At the client level, a new rpc is implemented to request a list of
events associated with the currently selected issue. In the UI, new left
and right buttons have been added to the details panel that allow users
to select next or previous event.

In the model, a new field is introduced to maintain the list of events.
The controller is switched from using the issue's sample event to using
this new field as the source of events. The issue's sample event is only
used for gutter icon generation.

Test: new and existing tests.
Bug: 294247767
Change-Id: Ic4fb6aff8b8bae1e7325b91baacc09a133544dd8