Reimplement AsyncDevicesGetter as DevicesService
The DevicesService consumes the flows of devices and templates from the
DeviceProvisioner, and produces a flow of Device objects (which wrap
DeviceProvisionerAndroidDevices).

LaunchCompatibility depends on IntelliJ state (the currently selected
run configuration, and dumb mode), so those are wrapped in Flows,
combined to produce the LaunchCompatibilityCheckerFlow, and that is then
combined with the devices flow to update devices when their launch
compatibility may have changed.

Bug: 299536545
Test: Included
Change-Id: Ia84104601af2d233b4e6fd33dd2034d655d794a4