Use execution scope to resolve compose files
In compose a project may have multiple modules that is referencing
different versions of the same dependency. In Studio the global
file index will have an entry for a source file from each of the
referenced versions.

In order to find the correct source file when "Going to source" we
should use the execution scope for the current run configuration.

Fixes: 307301775
Test: ModuleComposeResolverTest
Change-Id: If3016349ec9e3ac724e9994ad801aba2ee7b5a5d