Request project integrity sync for new android KMP plugin source sets
The `AndroidGradleProjectStartupActivity` validates the integrity of
Android models and models when a project is opened and triggers Gradle
sync in the case of any inconsistencies, which most likely are results
of previously misbehaving sync.

It was considered to filter the source sets by the known ones
defined on `IdeModuleWellKnownSourceSet`. However, this isn't exactly
consistent with the new Android KMP plugin which defines new ones
like: androidMain, androidInstrumentedTest, androidUnitTest but also
custom source set names.

For this reason, now, we are checking if the source set has a linked
androidFacet to include them in the validation integrity process.

Bug: 305944782
Test: OpenProjectIntegrationTest.testReopenProject_kmpWithAndroid
Change-Id: I2c799e84f48520e3fa025ddc17278f68e94923fb