Refresh TestAppResourceRepository children
TestAppResourceRepository can hold onto a reference to a linked module's
resource repository in some circumstances, and never refreshes when that
repository is disposed.

This change updates TestAppResourceRespository to refresh its child
repositories when roots change, akin to what AppResourceRepository does.

I haven't been able to directly reproduce a scenario that would lead to
the leak in b/304316363. The relevant facet that gets disposed is from
the androidTest module's linked main module; any time I close or remove
the main module, the androidTest module goes away as well. But I've
verified that this new invalidation logic does trigger whenever roots
change and works as expected, so since it's the same mechanism that
AppResourceRepository uses I'm hopeful it should address the issue.

Fixes: 304316363
Test: manually verified
Change-Id: Iaee225fa7feaad5482700eabfdeee933f31501e0