Use a project system token in UnusedResourcesProcessor
To handle resources programmatically generated by the build system,
where those can be found within build system files.

As well as improving modularity, this also addresses the problem of
not being sure whether a given file is a resource file, by allowing
the project system to take complete responsibility of files that it
owns -- and everything that is left, however improbable, should
actually be a resource file.

Bug: 147586218
Bug: 308400860
Test: existing
Change-Id: I8b7c5b8e03ef826f848b5c25cfd1764e342d5ef0