Use data context to get panel for Problems window actions
This makes use of the data context for actions in the Problems window to
avoid relying on the IssuePanelService.getSelectedPanel method (which is
then unnecessary).
It makes sense for the action context to know which panel it is linked
to, so we add that information to the context of
DesignerCommonIssuePanel.
In addition, this ensures that the panels are correctly updated with the
filter currently selected (as it is a project-wide setting and not a
per-panel setting).

Bug: N/A
Test: tests updated
Change-Id: Ife3edcfc29ec630e793ec1e3c09326607405270d