Add AGP output quickfixes for java language issues
... when target or source are not supported or deprecated.

When a module that has a target or source compatibility versions too
old, the compiler generates warnings and errors related to it that look
like these:

error: Source option 6 is no longer supported. Use 7 or later.
warning: [options] target value 7 is obsolete and will be removed in a future release

this change modifies AndroidGradlePluginOutputParser so it can now check
for known error messages and add quickfixes that point users to more
information or to the PSD, where they can change the source and target
compatibilities. If the used version is lower than 8 (and there is no
suggested version or it is lower or equal than 8) a quickfix to use
language level 8 in all modules is also shown.

Bug: 241067290
Test: AndroidGradlePluginOutputParserTest, JavaDeprecatedProviderTest
Change-Id: Ie609a650899eccb96a2b83fce641e57d0f16819f