Move pre-O Java/Kotlin Alloc results to capture stage
Currently, the pre-O Java/Kotlin Allocations feature presents the
allocation results in a table in the MainMemoryProfilerStage under
the streaming timeline and adjacent to the recording options. This
is the case for both recorded and imported recordings. This UX does
not make much sense given that other file based memory recordings
display their capture data in a capture stage (inconsident UX).

Thus, in this change, similar to the Native Allocations feature where
the user views the MainMemoryProfilerStage while recording, but on
termination of the recording can view the results in table form in a
capture stage, the pre-O Java/Kotlin Allocations will now behave the
same.

Bug: 277798936
Test: updated + added
Change-Id: I8677d46185502c1edd0119b5e7f57c981600d6fe