Reorder and comment file descriptions
Most of the dom.fileDescription/fileMetaData extensions are grouped and
ordered so that the ones handling the same types of XML files are listed
together, but a few are out of order. Fixing that and adding comments to
help differentiate.

This shouldn't have any effect on resolution since all of these
descriptors should be mutually exclusive. If there's a bug and they
aren't, it's possible the order could change - but it's already not
guaranteed (since none of them specify ordering constraints) and I've
kept the order consistent within each sub-grouping.

Change-Id: I1bc6f3d2f4652256b05918486e20cf02d7432436
Bug: n/a
Test: Covered by existing