Fix up handling/ownership of resource xml tags
This change ensures there's a full list of known root tags for res/xml
files, and that the various DomFileDescription classes don't overlap in
which tags they claim.

Pre-existing overlaps don't seem to have caused bugs, because the
overlaps only occurred between descriptions that claimed specific single
root tags and descriptions that claimed multiple tags. The single tag
descriptions have different handling in the platform and seem to
supersede the multi-tag descriptions, so the overlap didn't actually
matter. This change should clean things up and make the code more
understandable.

I'm not adding new tests, as there is a variety of good coverage over
these various types. (I know because I broke a lot of it while
refactoring and writing this change.) All of the representative types of
custom xml files have associated tests that are continuing to pass after
this change.

Bug: 290620639
Test: covered by existing
Change-Id: Ib7d96f0588acb1b36c2a338bdf8ec5a95a22ef4f