Add memory sample interval to config dialog
Initially, Native memory sampling interval was part of profiling tab.
In task-based ux, the profiling tab is removed and memory sampling
interval option is added along with existing cpu configs in edit
configurations dialog.

When task-based ux is enabled, the memory sampling interval config set
in this dialog is used and when disabled, the value from profiling tab
is used.

The existing PERFETTO_CONFIG is renamed to PERFETTO_SYSTEM_TRACE_CONFIG
& NATIVE_MEMORY_CONFIG renamed to PERFETTO_NATIVE_ALLOCATIONS_CONFIG to
include technology in name as to align with the naming of all children
of ProfilingConfiguration.

Bug: 277798411
Test: Added/Updated
Change-Id: I0bb6f902df3e65924b39515ef7a41d6d0e3027ce