Allow NlSupportedActions to be dynamic
This change is a refactoring allowing the actions to be dynamic and
change depending on the state of the surface.
The change should be no-op and just adds additional testing.

A follow-up CL will allow Compose to disable the actions for refresh and
toggle issue panel when interactive mode. This is to avoid the shortcuts
from conflicting with the interactive mode typing.

Bug: 310038272
Test: RefreshRenderActionTest, IssueNotificationActionTest
Change-Id: I2e5abc0776b6e0d99c751e7b9d56102d9c41b35a