Remove obsolete val from MergedManifestException
This exception takes `MergedManifestInfo` as a a parameter (which is
needed for an error message) and then stores is as a property (which is
not needed). Since the exception ends up being long-lived in the
`MergedManifestManager`, this can cause a memory leak since the info
object contains a reference to a `Facet` that can be disposed.

Removing the obsolete property removes the memory leak.

Fixes: 311397507
Test: n/a
Change-Id: I6873b2ff1e362c63939ff737b1fc9f69d9f1fc02