Function-typed variables and function literals can also be @Composable
This change adds the ability to return @Composable scope/expected
annotation holder for function-typed variables and function literals.
It also refactors the way this is computed so that the logic is
shared and only needs to be maintained in one place. I also added
copious comments to try to make it make sense, because it is very
hard to understand.

Bug: 309364913
Test: PsiUtilsTest.kt
Change-Id: Ibc19a6fb4a4156fd714d7daf66587ee31488f7b1