Define ExperimentalSettingsContributor
Redefine the ExperimentalSettingsConfigurable to be a
CompositeConfigurable, which uses this ExperimentalSettingsContributor
extension point to populate its contents.  Add a panel to the
ExperimentalSettingsConfigurable form to allow these contributed
configurables to be displayed.

At present there are no actual contributors: existing settings in this
panel will be migrated in subsequent changes, followed by existing
settings in sub-panels.

Bug: none filed
Test: existing
Change-Id: I325666659d28997918bfcc548220ff730a68f486