Disable AddComposableToFunctionQuickFix on containing function
This change disables the AddComposableToFunctionQuickFix when
it would be created in response to a COMPOSABLE_EXPECTED error.
With Kotlin 1.9.20, these apparently are generated on functions
that are not the correct element to be made @Composable. While
we figure out what's going on with that, we are disabling the
quick fix in the studio-hedgehog-dev branch.

Bug: b/309364913
Test: AddComposableToFunctionQuickFixTest.kt
Change-Id: Ieccb8d575f979105e22bf92ac7adc7d6731a4514