Add task grid for recording selection
The task grid component is slightly different for a recording
selection versus a process selection. As, unlike the process
based approach where all tasks are displayed whether supported
or not, the recording based approach only displays the supported
tasks.

The original TaskGrid composable code was refactored to share
much of the common code into a shared TaskGridContainer
composable.

Bug: 277798754
Test: added
Change-Id: Ibc9d29b3de227dacf244a2c88c9e5e38592322d2