Fallback for reading app theme when index is not ready
The app theme relies on the Manifest Indexer. For new projects, the
index might take a while to be ready and, in those cases, the initial
read might fail.

We had a fallback in places but it was not being triggered. This fixed
the code to have the fallback of reading directly from the manifest
merger while the index is pending to be built.

Test: ThemeUtilsIndexQueryTest, AndroidManifestIndex*Test
Fixes: 312085610
Change-Id: Iffe5f4163f6da362f7c30bd5c70478b6412d4032