Invalidate resources on all facets at first build
When resources are accessed before the first build (or sync) has been
done, some resource repositories/managers/etc may apparently contain
inaccurate data. To combat this, the ClearResourceCacheAfterFirstBuild
class exists and clears resources after that first build is complete.

Unfortunately, it only clears resources attached to the application
facet, not to all facets (ie, not for library facets). As part of
clearing the application facet, there's a chain of calls which clears
out the ResourceFolderRegistry, which manages
ResourceFolderRepositories. However, a library facet may still be
holding onto a (now defunct) ResourceFolderRepository. Since the library
facet isn't cleared, it retains that repository.

Later on, a change such as a file deletion happens in that folder.
ResourceFolderRegistry has a listener for that change, and sends it to
all the ReposourceFolderRepositories that it knows about. But since the
library has a dangling repository, that one is not updated. And voila,
havoc ensues.

I have manually verified the behavior in b/311198257 comment#2 is
fixed with this change. Unfortunately there's no existing test
infrastructure covering this specific logic. I'm getting this change in
now since it has a potentially large positive impact on a
difficult-to-repro bug, and I'd like to get as much feedback as
possible; I plan on following up with a test case.

Bug: 311198257
Fixes: 307329983
Test: Manually verified
Change-Id: I786f4f9ffe52845af28c0bd706ffa7303b777ede