Add task icon to task tab
...to serve as a visual indicator of what task the user is
currently on.

Note: It may seem redundant to have both the swing Icon mapping
added in this change in addition to the paths shown in the
instantiation of the task grid items in the same file. However,
this unfortunately can not be avoided as loading the icons in
Jewel via a swing Icon is not supported yet, and to use an icon
in an intellij toolwindow tab, we must use a swing Icon. A coming
refactor will reduce redundancy and unify the source of the icons.

Bug: 277797528
Test: manual
Change-Id: Ie7cd778926b1192bbd9a6588878fc013de50e7a3