Implement integration test for RunExternalSystemTaskAction
The RunExternalSystemTaskAction.perform is triggered when clicking
any task from the Gradle tool window ends up invoking
AndroidGradleTaskManager to run the task.

This test adds coverage for the entire flow since the user clicks
a specific task, which also covers some preparation execution validation
that happens in the platform before AndroidGradleTaskManager is
reached like the ones defined on LocalGradleExecutionAware.

Bug: 290572880
Test: GradleRunTaskActionIntegrationTest
Change-Id: I3018080c3596a40bac0ce2d8a7557a725abf2ef6