Only zoomToFit in UI Check when preview visibility changes
In UI Check, when errors are found or removed, it can affect the number
of visible previews if showing only previews with errors. If the
visibilty of some previews is modified, we zoom to fit.
However, we do not want to change the zoom level if no changes have
happened in the visibility of any preview.

Bug: 312538063
Test: UX change checked manually
Change-Id: Ifc03a2bd4a0545034b8eb211e6205efc77c087e8