Force refresh on resource change
This triggers a refresh when any resource changes. This will improve the
situation where the user changes a color in a drawable. Before this
change, the edit would not be reflected in the preview.
After this change, a refresh will be triggered.

This introduces a side effect where, if the user removes a resource or
updates and id, the preview might stop rendering (because of the
resource change). The user will need to refresh which is expected
anyway.

Bug: N/A
Test: New ResourceChangedListenerFlowTest
Change-Id: I75b49d7c12575c39db9bf26f10912026a7f2659f