Show loading screen in UI Check when compiling
When a full refresh is happening in UI Check mode, we currently see
previews updating and moving around as they become available, which is
not a great experience. This instead shows the UI Check empty panel
while the refreshing happens, so that the previews can all be displayed
at once when they are all ready.

Bug: 311068646
Test: UX change checked manually
Change-Id: Ib303de100fc52152ec013e2c7b0a5e920c7adb8e