Use K2 reference shortener for AddComposableAnnotationQuickFix
When AddComposableAnnotationQuickFix adds
`@androidx.compose.runtime.Composable` annotation, it does not shorten
the FQ name for K2. addAnnotation() util called by the quickfix calls
the reference shortener, but it seems the addAnnotation() internal sets
a wrong shortening target. It will be fixed from the upstream IJ.

This commit is a temporary workaround. See the more information in the
buganizer issue.

Bug: 311812857
Test: AddComposableAnnotationQuickFixTest#errorInPropertyInitializerWithAnonymousFunction_invokeOnFunctionCall
Change-Id: I24ab88179683c52d241014b94c40a8a75adc27bb