Attach debugger to sandbox sdk process.
The sandbox sdk process is a java process that doesn't come from the application (doesn't match the application's package name) but it somewhat tied to the application (similar name, and each application gets their own sandbox process). The sandbox process start command doesn't allow for waiting for the debugger on start-up so we need to attach the debugger retroactively.

Fixes: 313725396
Test: SandboxSdkLaunchTest
Change-Id: I08aa7ebadfe2032b0e73ebf52860b49b2fbae0cc