asdriver: invoke UI via performUserActivity()
This ensures that UI interactions are performed in a write-safe
context as needed (depending on the current modality state).
This is similar to how user actions are handled in production
(see IdeEventQueue or IdeKeyEventDispatcher).

For reference: TransactionGuardImpl.performActivity() sets
myWritingAllowed=true if the current ModalityState is write-safe.
Later, if the UI interaction spawns a new modal dialog,
TransactionGuardImpl.enteredModality() uses the current value of
myWritingAllowed to decide whether to mark the new ModalityState as
write-safe. Without this fix, all modal dialogs involved in the
test are incorrectly considered write-unsafe, leading to many
"Write-unsafe context!" errors in idea.log. For some reason,
these errors only started failing the test after updating
to IntelliJ 2023.3.

Potential follow-up work:

* Try to avoid using internal APIs in TransactionManagerImpl. The most
  promising alternative is to replace SwingUtilities.invokeLater()
  with Application.invokeLater() and just pass the appropriate
  ModalityState for the target component. However, it is unclear how
  to do this while still maintaining the property that 'finding' and
  'invoking' the component happen in the same EDT event.

* Try to avoid using SwingUtilities.invokeLater() since it does not
  handle the write-intent lock appropriately. The alternative way to
  run code on the EDT regardless of modality state is to use
  Application.invokeLater() with ModalityState.any().

* Consider whether we should only allow invoking UI elements that are
  within the topmost modal window (since in production these are the
  only elements that the user can interact with).

Bug: 314123372
Test: SamplesImportTest with IntelliJ 2023.3, 100 runs
Change-Id: I41d5c1c39c457c6db8bc9951800ae3f655c9b6a3