Ensure IDE and CLI Gradle callbacks for lint are consistent
This all started out with a simple bug report: lint doesn't flag
out-of-date *platform* dependencies -- as in

    implementation platform('androidx.compose:compose-bom:2023.08.00')

However, when I wrote a regression test for it, that was passing. It
turns out that it works from the CLI, but not in the IDE. The reason for
this discrepancy is that we have different code analysis backends for
Groovy handling in those two scenarios. In the IDE, we're using
IntelliJ's Groovy PSI, and from the CLI, we're using Groovy itself.
We've built a simple abstraction on top of these parsetrees -- there are
callbacks to GradleScanners where you are told about "properties" and
"method calls". For example, whether you have

    android {
       defaultConfig {
            minSdkVersion 23
        }
    }
or

    android.defaultConfig.minSdkVersion = 23

in both cases you're told about the property "minSdkVersion" with parent
"defaultConfig" and grand parent "android", as a property.

Well, it turns out this mapping layer was quite different between the
two. And that seems bad. 90% of the Grooy related tests are on the CLI
side, so they may not actually work in the IDE.

So to make sure that Gradle detectors are behaving the same way whether
from the command line in the IDE, there's now a new set of tests which
are "pretty printing" the exact callback history for a set of Gradle
scripts -- using language features like closures, variable declarations,
named parameters, and so on.

Of course, initially the behaviors did *not* match. So this CL also
fixes this. At first it looked like it was mainly the IDE visitor that
needed fixing -- it had received fewer improvements than on the Groovy
side one. But when I was digging into one of the last diffs between the
two, I noticed that what the CLI was doing was suspicious --it was
actually treating nested calls (`foo(bar("hello"))`) as if they are
nested properties (`foo.bar="hello"`) -- that's not right. So this CL
also fixes the Groovy visitor to not do that.

(Future work: Right now, the tests are duplicated, one on each side,
with the same golden-strings and callback pretty-printer; in the future
we could attempt to share testing resources between the CLI and IDE lint
tests and unify this.)

Bug: Not filed (but related to 313937481)
Test: Unit tests included
Change-Id: I0fc9bdca632e476156f2f41516cf12205985c3de