Move NdkFacet to Gradle project system module
This involves taking a dependency on the Gradle project system from
the gradle-specific native-symbolizer extension module, which is fine,
and from the android-navigation module implementing the Android
Project View, which is less fine.  However, the android-navigation
module already depends on the Gradle "IDE models", which suggests that
it is already not cleanly abstracted; this change doesn't actively
make that worse, so reducing or removing those gradle-based
dependencies will be handled in further work.

Bug: 147586218
Test: existing
Change-Id: Iaad589a4d0da053525d63fd1657c8194bb2b0cd4